Fix behaviour of DraftPublished and ChangeMerged
Draft published event should not implement RepositoryModifiedEvent since it
does not really modify the repo.

Change Merged has the potential the modify the repo so we must implement
RepositoryModifiedEvent.

[JENKINS-25047]
